+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ## 0.0.4
+  - Re-syncs with Elasticsearch 8.11 [#91](https://github.com/elastic/logstash-filter-elastic_integration/pull/91)
   - Restores support for `redact` processor using its x-pack licensed implementation [#90](https://github.com/elastic/logstash-filter-elastic_integration/issues/90)
 
 ## 0.0.3